requirements.txt
aiofile==3.9.0
alabaster==1.0.0
alembic==1.16.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.2
asttokens==3.0.0
attr==0.3.2
attrs==26.1.0
Authlib==1.6.9
av==15.1.0
babel==2.17.0
backcall==0.2.0
backports.tarfile==1.2.0
bcrypt==4.3.0
beartype==0.22.9
beautifulsoup4==4.14.3
blinker==1.9.0
boltons==25.0.0
brotli==1.2.0
cachetools==5.5.2
caio==0.9.25
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configparser==7.1.0
contextlib2==21.6.0
contourpy==1.3.3
cryptography==43.0.1
ctranslate2==4.6.0
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==4.10.2
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.21.2
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.0
fastapi==0.133.0
faster-whisper==1.2.0
fastmcp==3.2.3
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
flatbuffers==25.2.10
fonttools==4.59.1
frozendict==2.4.7
fsspec==2025.9.0
funcsigs==1.0.2
future==1.0.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.47.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gradio==6.12.0
gradio_client==2.4.1
greenlet==3.2.3
groovy==0.1.2
groq==0.31.0
gunicorn==23.0.0
h11==0.16.0
hf-gradio==0.3.0
HTMLParser==0.0.2
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.3
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
jupyterlab_widgets==3.0.14
keyring==25.7.0
kiwisolver==1.4.9
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.27.0
mdurl==0.1.2
mock==5.1.0
more-itertools==10.6.0
mpmath==1.3.0
multitasking==0.0.12
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==2.31.0
openapi-pydantic==0.5.1
openenv-core==0.2.3
openpyxl==3.1.5
opentelemetry-api==1.41.0
orjson==3.11.8
packaging==24.2
pandas==2.3.2
parso==0.8.4
pathable==0.5.0
peewee==4.0.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.9.2
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.28.2
pure_eval==0.2.3
py-key-value-aio==0.4.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.13.1
pydantic_core==2.33.2
pydub==0.25.1
pyenv-win==3.1.1
Pygments==2.19.1
pyjnius==1.6.1
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.3
pyperclip==1.11.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
railroad==0.5.0
redis==5.1.0
referencing==0.37.0
requests==2.32.2
requests-oauthlib==2.0.0
rich==14.3.4
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
sse-starlette==3.3.4
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
thread==2.0.5
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.4.1
tomli_w==1.2.0
tomlkit==0.14.0
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uncalled-for==0.3.1
uritemplate==4.2.0
urllib3==2.3.0
urllib3-secure-extra==0.1.0
uv==0.11.6
uvicorn==0.41.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yfinance==1.2.0
zipp==3.21.0
